+++ sources/src/repositories/history.repository.ts
@@ -16,14 +16,14 @@ import { EntityUniqueConflictError } from "../errors";
 import { HistoryEntity, HistoryEntityRelations } from "../models";
 
 export interface HistoryOptions extends Options {
-    crud?: true;
+    history?: true;
     maxDate?: Date;
 }
 
 /**
  * Repository Type
  */
-export interface HistoryCrudRepository<
+export interface HistoryRepository<
     Model extends HistoryEntity,
     ModelRelations extends HistoryEntityRelations
 > extends DefaultCrudRepository<Model, string, ModelRelations> {}
@@ -31,7 +31,7 @@ export interface HistoryCrudRepository<
 /**
  * Repository Mixin
  */
-export function HistoryCrudRepositoryMixin<
+export function HistoryRepositoryMixin<
     Model extends HistoryEntity,
     ModelRelations extends HistoryEntityRelations
 >() {
@@ -46,7 +46,7 @@ export function HistoryCrudRepositoryMixin<
         >
     >(
         superClass?: RepositoryClass
-    ): RepositoryClass & Class<HistoryCrudRepository<Model, ModelRelations>> {
+    ): RepositoryClass & Class<HistoryRepository<Model, ModelRelations>> {
         const parentClass: Class<DefaultCrudRepository<
             Model,
             string,
@@ -54,7 +54,7 @@ export function HistoryCrudRepositoryMixin<
         >> = superClass || DefaultCrudRepository;
 
         class Repository extends parentClass
-            implements HistoryCrudRepository<Model, ModelRelations> {
+            implements HistoryRepository<Model, ModelRelations> {
             constructor(ctor: Ctor<Model>, dataSource: juggler.DataSource) {
                 super(ctor, dataSource);
             }
@@ -165,7 +165,7 @@ export function HistoryCrudRepositoryMixin<
                 entity: DataObject<Model>,
                 options?: HistoryOptions
             ): Promise<Model> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.create(entity, options);
                 }
 
@@ -178,7 +178,7 @@ export function HistoryCrudRepositoryMixin<
                 entities: DataObject<Model>[],
                 options?: HistoryOptions
             ): Promise<Model[]> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.createAll(entities, options);
                 }
 
@@ -227,7 +227,7 @@ export function HistoryCrudRepositoryMixin<
                 filter?: Filter<Model>,
                 options?: HistoryOptions
             ): Promise<(Model & ModelRelations)[]> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.find(filter, options);
                 }
 
@@ -261,7 +261,7 @@ export function HistoryCrudRepositoryMixin<
                 filter?: Filter<Model>,
                 options?: HistoryOptions
             ): Promise<(Model & ModelRelations) | null> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.findOne(filter, options);
                 }
 
@@ -301,7 +301,7 @@ export function HistoryCrudRepositoryMixin<
                 filter?: Filter<Model>,
                 options?: HistoryOptions
             ): Promise<Model & ModelRelations> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.findById(id, filter, options);
                 }
 
@@ -349,7 +349,7 @@ export function HistoryCrudRepositoryMixin<
                 where?: Where<Model>,
                 options?: HistoryOptions
             ): Promise<Count> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.count(where, options);
                 }
 
@@ -385,7 +385,7 @@ export function HistoryCrudRepositoryMixin<
                 id: string,
                 options?: HistoryOptions
             ): Promise<boolean> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.exists(id, options);
                 }
 
@@ -519,7 +519,7 @@ export function HistoryCrudRepositoryMixin<
                 entity: Model,
                 options?: HistoryOptions
             ): Promise<void> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.update(entity, options);
                 }
 
@@ -539,7 +539,7 @@ export function HistoryCrudRepositoryMixin<
                 where?: Where<Model>,
                 options?: HistoryOptions
             ): Promise<Count> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.updateAll(data, where, options);
                 }
 
@@ -566,7 +566,7 @@ export function HistoryCrudRepositoryMixin<
                 data: DataObject<Model>,
                 options?: HistoryOptions
             ): Promise<void> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.updateById(id, data, options);
                 }
 
@@ -586,7 +586,7 @@ export function HistoryCrudRepositoryMixin<
                 data: DataObject<Model>,
                 options?: HistoryOptions
             ): Promise<void> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.replaceById(id, data, options);
                 }
 
@@ -623,7 +623,7 @@ export function HistoryCrudRepositoryMixin<
                 entity: Model,
                 options?: HistoryOptions
             ): Promise<void> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.delete(entity, options);
                 }
 
@@ -640,7 +640,7 @@ export function HistoryCrudRepositoryMixin<
                 where?: Where<Model>,
                 options?: HistoryOptions
             ): Promise<Count> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.deleteAll(where, options);
                 }
 
@@ -659,7 +659,7 @@ export function HistoryCrudRepositoryMixin<
                 id: string,
                 options?: HistoryOptions
             ): Promise<void> {
-                if (options && options.crud) {
+                if (options && options.history) {
                     return super.deleteById(id, options);
                 }